Car-Parts-Bot/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.35
PyMySQL==1.1.1
python-dotenv==1.0.1
pydantic==2.9.2
requests==2.32.3
openai==2.8.1
uvicorn==0.32.0
gunicorn==23.0.0
playwright==1.47.0
deep-translator==1.11.4
# langdetect==1.0.9
pytesseract==0.3.13
Pillow==10.4.0
 

